refactor(project-files-view): extract preview owner
